requirements.txt
# force vercel rebuild
flask>=3.1.2
prometheus-client>=0.23.0
flask-cors>=6.0.1
flask-limiter>=3.13
redis>=7.0.1
requests>=2.32.5
pydantic>=2.11.9
python-dotenv>=1.1.1
loguru>=0.7.3
itsdangerous>=2.2.0
pycountry>=24.6.1
google-genai>=1.49.0
azure-cognitiveservices-speech>=1.41.1
